Revert "[FFI][RUNTIME] Introduce runtime boxed types for int/float/bool (#16183)"
This reverts commit 5f22be4d83ca698e316ac342f32f5b4d38155ca8.

+++ vta/python/vta/transform.py
@@ -19,7 +19,6 @@
 import tvm
 from tvm import te
 from tvm.topi import utils
-from tvm.script import tir as T
 
 from .environment import get_env
 
@@ -1047,19 +1046,19 @@ def InjectALUIntrin():
                 assert len(dst_coeff) > 1
                 assert len(extents) != 0
                 tvm.ir.assert_structural_equal(
-                    analyzer.simplify(idxm(src_coeff[-1], env.BATCH * env.BLOCK_OUT)), T.int32(0)
+                    analyzer.simplify(idxm(src_coeff[-1], env.BATCH * env.BLOCK_OUT)), 0
                 )
                 tvm.ir.assert_structural_equal(
-                    analyzer.simplify(idxm(dst_coeff[-1], env.BATCH * env.BLOCK_OUT)), T.int32(0)
+                    analyzer.simplify(idxm(dst_coeff[-1], env.BATCH * env.BLOCK_OUT)), 0
                 )
-                tvm.ir.assert_structural_equal(src_coeff[-2], T.int32(1))
-                tvm.ir.assert_structural_equal(dst_coeff[-2], T.int32(1))
+                tvm.ir.assert_structural_equal(src_coeff[-2], 1)
+                tvm.ir.assert_structural_equal(dst_coeff[-2], 1)
                 if env.BATCH > 1:
                     assert len(src_coeff) > 2
                     assert len(dst_coeff) > 2
                     assert len(extents) > 1
-                    tvm.ir.assert_structural_equal(src_coeff[-3], T.int32(env.BLOCK_OUT))
-                    tvm.ir.assert_structural_equal(dst_coeff[-3], T.int32(env.BLOCK_OUT))
+                    tvm.ir.assert_structural_equal(src_coeff[-3], env.BLOCK_OUT)
+                    tvm.ir.assert_structural_equal(dst_coeff[-3], env.BLOCK_OUT)
 
                 # Apply tensorization of the loop coefficients
                 src_offset = src_coeff[-1]
